[d3d8] Refactor D3D8Vertex/IndexBuffer implementation
