+++ src/d3d8/meson.build
@@ -1,6 +1,7 @@
 d3d8_res = wrc_generator.process('version.rc')
 
 d3d8_src = [
+  'd3d8_buffer.cpp',
   'd3d8_device.cpp',
   'd3d8_interface.cpp',
   'd3d8_main.cpp',
